% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Build_Matrix.R
\name{build_matrix}
\alias{build_matrix}
\title{Constructing a matrix}
\usage{
build_matrix(psInputFile, pbLog = FALSE)
}
\description{
Storing the variance and covariance in a list
}
\details{
---
title: Constructing multivariate estimates from analyses by parts in a matrix
date:  "`r Sys.Date()`"
---
}
