% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakePD.R
\name{check_transform_positivedefinit}
\alias{check_transform_positivedefinit}
\title{Bending of Matrix for Different Random Effects}
\usage{
check_transform_positivedefinit(psInputFile, psOptionRatio, psRatio,
  pbLog = FALSE)
}
\description{
Bending of Matrix for Different Random Effects
}
