% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateVarCovarMix99.R
\name{create_parameter_varCovar_mix99}
\alias{create_parameter_varCovar_mix99}
\title{Build Parameter with Variances and Covariances for Mix99}
\usage{
create_parameter_varCovar_mix99(psInputFile, psOutputFile, pbLog = FALSE)
}
\description{
Build Parameter with Variances and Covariances for Mix99
}
