% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakePD.R
\name{make_pd_rat_ev}
\alias{make_pd_rat_ev}
\title{Bending of Matrix A with Ratio}
\usage{
make_pd_rat_ev(A, pn_max_ratio)
}
\arguments{
\item{A}{input matrix}

\item{pn_max_ratio}{max ratio}
}
\value{
Bended positive-definite matrix A with ratio
}
\description{
## Parametrisation
The current implementation bends the given matrix in a fixed fashion. All it does is to increment the negative
eigenvalues between 0 and the smallest positive eigenvalue. If the smallest positive eigenvalue is very small, then
the corrected eigenvalues will be about 100 times smaller which corresponds to the hard-coded factor in the computation
of the weight `wr`.

For some applications it might be interesting to be able to specify a ratio between the largest and the smallest
eigenvalue. With that it should be possible to scale all eigenvalues to be within that ratio boundary.

## Ratio of Eigenvalues
Instead of just changing the negative eigenvalues by mapping them between zero and the smallest eigenvalue,
we would like to correct eigenvalues such that all of them are poitive and such that the ratio between the
largest and the smallest eigenvalue is below a certain threshold.

In principle this approach can be implemented similarly to the one that corrects the negative eigenvalues,
presented so far. The only adaptation that one must do is to replace the limit of the eigenvalues that must
be changed from 0 to that number determined by the maximum ration. Everything else should stay the same,
modulo a special treatment of any negative eigenvalues. An easy solution to that might be to do it in two
steps, first make all eigenvalues positive using an approach used in `makePD2()`. Then we can use a second
step to come up with a correction that ensures a maximum ration between eigenvalues. On the other hand it
should be possible to provide a one step solution.

The function is called `make_pd_rat_ev()`. As arguments the function takes the input matrix and a maximum
ratio between largest and smallest eigenvalue.
}
