% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreatePlot.R
\name{plot_gencorr}
\alias{plot_gencorr}
\title{Create Plot to Overview VCE results for genetic correlation}
\usage{
plot_gencorr(psInputFile)
}
\description{
Create Plot to Overview VCE results for genetic correlation
}
