% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadVCE.R
\name{read_vce}
\alias{read_vce}
\title{Reading VCE results}
\usage{
read_vce(psInputFile, pbLog = FALSE)
}
\description{
First attempt to read variance component estimation (VCE) from a csv-file
and storing the input in tibble
}
